Move FlowTrace,FlowTraceHop next to FlowHistory
