+++ projects/batfish/src/main/java/org/batfish/symbolic/smt/CounterExample.java
@@ -26,8 +26,6 @@ import org.batfish.datamodel.Edge;
 import org.batfish.datamodel.FilterResult;
 import org.batfish.datamodel.Flow;
 import org.batfish.datamodel.FlowDisposition;
-import org.batfish.datamodel.FlowTrace;
-import org.batfish.datamodel.FlowTraceHop;
 import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.Ip;
 import org.batfish.datamodel.IpAccessList;
@@ -43,6 +41,8 @@ import org.batfish.symbolic.Graph;
 import org.batfish.symbolic.GraphEdge;
 import org.batfish.symbolic.Protocol;
 import org.batfish.symbolic.answers.FlowHistory;
+import org.batfish.symbolic.answers.FlowTrace;
+import org.batfish.symbolic.answers.FlowTraceHop;
 import org.batfish.symbolic.utils.Tuple;
 
 class CounterExample {
